requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
openai
sentence-transformers
faiss-cpu
openai-agents
pydantic
dotenv